change cts type to int, save some space
Originally committed as revision 18830 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavformat/movenc.c
@@ -49,7 +49,7 @@ typedef struct MOVIentry {
     unsigned int samplesInChunk;
     char         key_frame;
     unsigned int entries;
-    int64_t      cts;
+    int          cts;
     int64_t      dts;
 } MOVIentry;
